Move sockets out from communication into separate class
