+++ frosthaven_assistant/lib/services/service_locator.dart
@@ -1,5 +1,6 @@
 import 'package:frosthaven_assistant/Resource/state/game_state.dart';
 import 'package:frosthaven_assistant/Resource/settings.dart';
+import 'package:frosthaven_assistant/services/network/connection.dart';
 import 'package:get_it/get_it.dart';
 
 import 'network/communication.dart';
@@ -12,4 +13,5 @@ void setupGetIt() {
   getIt.registerLazySingleton<GameState>(() => GameState());
   getIt.registerLazySingleton<Communication>(() => Communication());
   getIt.registerLazySingleton<Network>(() => Network());
+  getIt.registerLazySingleton<Connection>(() => Connection());
 }
